+++ tutorials/bounding-box-crop/src/components/CustomHandle.tsx
