+++ tutorials/bounding-box-crop/src/components/ErrorAlert.tsx
