+++ tutorials/bounding-box-crop/src/components/Spinner.tsx
